Move generic functions to `Utils.ts`
